feat: added styling to plans cards on Pricing page
